chore: rename core dir to domain
