+++ server-new/src/setup/graphql/schema.ts
@@ -1,8 +1,8 @@
 import { builder } from './builder';
-import * as post from '~/core/example/post/post.resolver';
-import * as user from '~/core/user/user.resolver';
-import * as session from '~/core/session/session.resolver';
-import * as chat from '~/core/example/chat/chat.resolver';
+import * as post from '~/domain/example/post/post.resolver';
+import * as user from '~/domain/user/user.resolver';
+import * as session from '~/domain/session/session.resolver';
+import * as chat from '~/domain/example/chat/chat.resolver';
 
 export function setupSchema() {
   /**
