+++ server-new/src/utils/auth.ts
@@ -3,8 +3,8 @@ import { DrizzlePostgreSQLAdapter } from '@lucia-auth/adapter-drizzle';
 
 import { config } from './config';
 import { getDb } from '~/db';
-import { sessionTable } from '~/core/session/session.db';
-import { userTable } from '~/core/user/user.db';
+import { sessionTable } from '~/domain/session/session.db';
+import { userTable } from '~/domain/user/user.db';
 
 let lucia: Lucia | null = null;
